pyproject.toml
[tool.poetry]
name = "nomad-pyteam-backend"
version = "0.1.0"
description = ""
authors = ["DaehunGwak <rhkrgns56@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
